pyproject.toml
[tool.poetry]
name = "codecampapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.8"}
psycopg = {extras = ["binary"], version = "^3.2.4"}
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.7.1"
alembic = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88